分离defunct
